Create Project 1 Task B
Write a program maze_runner.cpp that reads from the cin a sequence of one or more letters and treat it as a list of directions that take you through the maze array to see whether or not you exit the maze at the correct location or not. Space characters can be anywhere in the input. After the input ends (end-of-file is reached), the program should compute and print the result of the input summation.